googlenet.m
Special funtion has been implemented: findLayersToReplace(lgraph). It automatically replaces name of 2 layers. We could do this on our own but if such a function exists we should use it :) 
In most networks, the last layer with learnable weights is a fully connected layer. Replace this fully connected layer with a new fully connected layer with the number of outputs equal to the number of classes in the new data set (in our case - 2).